grsec: Apply grsecurity-2.2.2-2.6.32.57-201202161839.patch
commit fec67d63c16452e7a59f5952972e49d883391ab1
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Feb 16 18:38:32 2012 -0500

    Update configuration help and name for GRKERNSEC_PROC_MEMMAP

 grsecurity/Kconfig |   10 ++++++----
 1 files changed, 6 insertions(+), 4 deletions(-)

commit c5ac1c8fdb233ee524d8449c31cb73abacfd3162
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Feb 16 18:18:01 2012 -0500

    optimize the check a bit

 fs/exec.c |    4 ++--
 1 files changed, 2 insertions(+), 2 deletions(-)

commit 494cb302690cba6daf117583e405645dd40c5364
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Feb 16 18:00:45 2012 -0500

    smile VUPEN :D
    (limit argv+env to 1MB for suid/sgid binaries)

 fs/exec.c |   11 +++++++++++
 1 files changed, 11 insertions(+), 0 deletions(-)

commit f44d1be39ce6336f8270bff0040babf09c040ef7
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Feb 16 17:49:33 2012 -0500

    Address Space Protection -> Memory Protections (suggested on IRC for consistency)

 grsecurity/Kconfig |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

commit 11749696a05a2a4090d036d8c51a99acdea4a0d5
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu Feb 16 17:45:06 2012 -0500

    Change the long long type for exec_id to the proper u64

 include/linux/sched.h    |    2 +-
 include/linux/seq_file.h |    2 +-
 2 files changed, 2 insertions(+), 2 deletions(-)

Signature-tree: 1ec9038142a3350d796d1fb3945b375b5a7c37e2

+++ include/linux/seq_file.h
@@ -24,7 +24,7 @@ struct seq_file {
 	struct mutex lock;
 	const struct seq_operations *op;
 #ifdef CONFIG_GRKERNSEC_PROC_MEMMAP
-	long long exec_id;
+	u64 exec_id;
 #endif
 	void *private;
 };
